Changes from review. Refactored all access logic into a new class
