+++ conf/ds-image-behaviour.yaml
@@ -24,3 +24,11 @@ config:
   userhome: "Home sweet ${user.home}"
   # Fallbacks for system properties can be defined with :-
   fallback: ${nonexisting.property:-87}
+  #Url to licensemodule
+  licensemodule:
+    url: http://localhost:10001/ds-license/v1/
+    
+  # Default images returned if there is no access to image or image(id) does not exist
+  images:
+    no_access: no_access.png
+    non_existing: non_existing.png
